refactor: isEncoded함수 제거 [AIR-125]
