+++ aircnc/src/test/java/com/gurudev/aircnc/domain/member/entity/PasswordTest.java
@@ -2,8 +2,10 @@ package com.gurudev.aircnc.domain.member.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
+import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 
-import com.gurudev.aircnc.configuration.PasswordEncryptor;
+import com.gurudev.aircnc.infrastructure.security.PasswordEncryptor;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
@@ -28,11 +30,30 @@ class PasswordTest {
 
   @ParameterizedTest
   @CsvSource(value = {"12345678", "123456789012345"})
-  void 비밀번호_암호화_테스트(String rawPassword) {
+  void 비밀번호_암호화_성공_테스트(String rawPassword) {
     Password password = new Password(rawPassword);
 
     password.encode(passwordEncryptor);
 
-    assertThat(password.isEncoded()).isTrue();
+    assertThat(password.matches(passwordEncryptor, new Password(rawPassword))).isTrue();
+  }
+
+  @ParameterizedTest
+  @CsvSource(value = {"12345678", "123456789012345"})
+  void 암호화된_비밀번호_불일치_테스트(String rawPassword) {
+    Password password = new Password(rawPassword);
+
+    password.encode(passwordEncryptor);
+
+    assertThat(password.matches(passwordEncryptor, new Password("invalidPassword"))).isFalse();
+  }
+
+  @Test
+  void 암호화_되지_않은_비밀번호는_일치여부_확인불가() {
+    Password password = new Password("abcdefggg");
+
+    assertThatIllegalStateException().isThrownBy(
+        () -> password.matches(passwordEncryptor, new Password("abcdefggg"))
+    );
   }
 }
